+++ app/Http/Controllers/NoteController.php
@@ -80,51 +80,70 @@ class NoteController extends Controller
     // Display instance o a model update form
     public function edit(Note $note)
     {
-        return view('notes.edit', ['note' => $note]);
+        if(($note->user_id == auth()->user()->id) || (auth()->user()->is_admin)){
+            return view('notes.edit', ['note' => $note]);
+        }
+        else{
+            return back()->with('message', 'Nie posiadasz uprawnień!');
+        }
+
+        
     }
 
     // Validate data from creation form and update instance in database
     public function update(Request $request, Note $note)
     {
-        //Validating request from form
-        $validated = $request->validate([
-            'subject_id' => 'numeric',
-            'name' => 'string',
-            'date' => 'date',
-            'description' => 'string',
+        if(($note->user_id == auth()->user()->id) || (auth()->user()->is_admin)){
+            //Validating request from form
+            $validated = $request->validate([
+                'subject_id' => 'numeric',
+                'name' => 'string',
+                'date' => 'date',
+                'description' => 'string',
 
-        ]);
+            ]);
 
-        //Changing data only if request had new data
-        if($request->has('subject_id')){
-            $note->subject_id = $validated['subject_id'];
-        }
-        if($request->has('name')){
-            $note->name = $validated['name'];
-        }
-        if($request->has('date')){
-            // Additional validation for time inputs
-            $date = $this->validateTime($validated['date']);
-            if($date==false)
-            {
-                return back()->withErrors(['start' => 'Invalid date in form field']);
+            //Changing data only if request had new data
+            if($request->has('subject_id')){
+                $note->subject_id = $validated['subject_id'];
             }
-            $note->date = $date;
-        }
-        if($request->has('description')){
-            $note->description = $validated['description'];
-        }
-        $note->save();
+            if($request->has('name')){
+                $note->name = $validated['name'];
+            }
+            if($request->has('date')){
+                // Additional validation for time inputs
+                $date = $this->validateTime($validated['date']);
+                if($date==false)
+                {
+                    return back()->withErrors(['start' => 'Invalid date in form field']);
+                }
+                $note->date = $date;
+            }
+            if($request->has('description')){
+                $note->description = $validated['description'];
+            }
+            $note->save();
 
-        return redirect('/notes/'.$note->id)->with('message', 'Edycja pomyślna!');
+            return redirect('/notes/'.$note->id)->with('message', 'Edycja pomyślna!');
+        }
+        else{
+            return back()->with('message', 'Nie posiadasz uprawnień!');
+        }
+        
             
     }
 
     // Soft delete instance of a model
     public function destroy(Note $note)
     {
-        $note->delete();
-        return redirect('/')->with('message', 'Usunięto!');
+        if(($note->user_id == auth()->user()->id) || (auth()->user()->is_admin)){
+            $note->delete();
+            return redirect('/')->with('message', 'Usunięto!');
+        }
+        else{
+            return back()->with('message', 'Nie posiadasz uprawnień!');
+        }
+        
     }
 
     // // Restore trashed (soft deleted) instance of a model
